litespeed dropped
Out With Old, In With The New
REST API should be completely decoupled from the old litespeed stuff
API only works on local machine, (and REST acts as a web-server)  so I need to see if I can make a wrapper from the OpenAPI doc generated, or if I need to roll my own using requests.

Alot of old code is being dropped; I've source controlled it so theres no reason to keep it. If i need it I can just dig through the git history.

I was thinking I'd use flask or django to build the webside of things; but I don't think I need to; I could use fastapi to route my traffic (and disable the docs). HTML and Files can be served directly via responses.